Fix lint issues in actions
